+++ app/Modules/Clients/Views/clients/index.blade.php
@@ -56,7 +56,7 @@
                 <div class="box box-primary">
 
                     <div class="box-body no-padding">
-                        {!! $dataTable->table() !!}
+                        {!! $dataTable->table(['class' => 'table dt-responsive display', 'width' => '100%', 'cellspacing' => '0']) !!}
                     </div>
 
                 </div>