+++ app/Modules/Scheduler/Views/categories/_dataTable.blade.php
@@ -6,8 +6,8 @@
     {{--<script src="/assets/plugins/datatables.net-buttons/js/dataTables.buttons.min.js"></script>--}}
     {{--<script src="/vendor/datatables/buttons.server-side.js"></script>--}}
     {!! $dataTable->scripts() !!}
-    <script>
+   {{-- <script>
         var htmlstr = '<input type="checkbox" id="bulk-select-all"/> ';
         $('.bulk-record').html(htmlstr)
-    </script>
+    </script>--}}
 @endpush