+++ app/Modules/TimeTracking/Views/time_tracking/project_index.blade.php
@@ -124,7 +124,7 @@
 
                         {{--</table>--}}
 
-                        {!! $dataTable->table() !!}
+                        {!! $dataTable->table(['class' => 'table dt-responsive display', 'width' => '100%', 'cellspacing' => '0']) !!}
 
                     </div>