+++ app/Modules/Users/Views/users/index.blade.php
@@ -50,7 +50,7 @@
 
                     <div class="box-body">
 
-                        {!! $dataTable->table() !!}
+                        {!! $dataTable->table(['class' => 'table dt-responsive display', 'width' => '100%', 'cellspacing' => '0']) !!}
 
                     </div>
                 </div>